refactor: 내 정보 관련 쿼리 서비스 분리
